+++ src/leaflet/overlay/graphic/index.js
@@ -1,5 +1,8 @@
+import {CloverStyle, cloverStyle} from './CloverStyle';
 import {CircleStyle, circleStyle} from './CircleStyle';
 import {Graphic, graphic} from './Graphic';
 
+
+export {CloverStyle, cloverStyle} ;
 export {CircleStyle, circleStyle} ;
 export {Graphic, graphic} ;